+++ packages/chatkit/src/index.ts
@@ -1,5 +1,6 @@
 export * from './interrupt.js';
 export * from './message.js';
 export * from './options.js';
+export * from './pet.js';
 export * from './chatkit.js';
 export * from './commands.js';