+++ examples/managed-chatkit-angular/frontend/CHANGELOG.md
@@ -1,5 +1,19 @@
 # managed-chatkit-angular-frontend
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [5c0cab1]
+- Updated dependencies [f594038]
+- Updated dependencies [361c358]
+- Updated dependencies
+- Updated dependencies [96aac52]
+- Updated dependencies [169d5d5]
+  - @xpert-ai/chatkit-types@0.3.0
+  - @xpert-ai/chatkit-angular@0.2.0
+  - @xpert-ai/chatkit-web-component@0.3.0
+
 ## 0.0.8
 
 ### Patch Changes